+++ glue/viewers/matplotlib/__init__.py
@@ -1,0 +1,1 @@
+from . import toolbar_mode  # noqa