+++ glue/viewers/profile/qt/profile_tools.py
@@ -16,7 +16,7 @@ from glue.utils.qt import load_ui, fix_tab_widget_fontsize
 from glue.viewers.profile.qt.mouse_mode import NavigateMouseMode, RangeMouseMode
 from glue.core.qt.fitters import FitSettingsWidget
 from glue.utils.qt import Worker
-from glue.viewers.common.qt.tool import Tool
+from glue.viewers.common.tool import Tool
 from glue.viewers.image.state import AggregateSlice
 from glue.core.aggregate import mom1, mom2
 from glue.core import BaseData, Subset